% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCall.R
\name{cpCall}
\alias{cpCall}
\title{Title}
\usage{
cpCall(
  bedFile,
  bamFilesTumors,
  bamFilesNormals = NULL,
  normalization = "median_gene",
  lstGenes = NULL,
  purity = NULL,
  selectNormalsByCosine = FALSE,
  numOfNormalsByCosine = 3,
  tumorsToUseAsNormal = 3,
  FFPE = F,
  smooth = FALSE,
  callOnly = FALSE,
  ccObj = NULL
)
}
\arguments{
\item{ccObj}{}
}
\description{
A short description...
}
